requirements.txt
Django==2.2.10
django-crispy-forms==1.8.0
django-summernote==0.8.11.4
django-taggit==1.1.0
html5lib==1.0.1
Pillow==6.2.1
PyPDF2==1.26.0
Pygments==2.5.2
python-decouple==3.2
pytz==2019.3
reportlab==3.5.32
six==1.13.0
sqlparse==0.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3
Markdown==3.1.1
Unidecode==1.1.1
markdown-checklist==0.4.1
django-cryptography==1.0
